changed movement to CHARACHTER CONTROLLER component. Allows for easier use, but is more simple
